disable CONFIG_SITE envvar at build time
CONFIG_SITE is an environment variable which may point to a global autoconf initialization script,
which is run at the start of every autoconf invocation and can modify its parameters.
Most systems don't have that script, and don't set that variable. On OpenSUSE, the variable is set.
The default script overwrites paths to local builds of libraries from `lib` to `lib64`, breaking our
builds. This override makes no sense for our project, since all libraries are linked statically, and
only ever built in 64-bit versions (the script allows side by side 32- and 64-bit builds).

It is unlikely that this script is used in practice, since the possibility of breaking random builds is
too high. It's also relevant only for building C code which relies on autoconf. Most new projects and
projects in modern languages don't do that.

For this reason I unset CONFIG_SITE in our default builds of Tarantool.