fix: avoid no such user error when the user is dropped during DROP USER IF EXISTS execution
Here is how DROP USER IF EXISTS query can result in such error:

- [client 1]: check user exists
- [client 1]: send CAS request 1
- [client 2]: check user exists
- [client 2]: send CAS request 2
- [leader]: recv CAS request 1 -> access_check: ok -> check predicate: ok -> apply(async)
- [leader]: recv CAS request 2 -> access_check: no such user
                               /|\
                                |
                                |
                             applied

To address such errors, access_check was made to not return errors when
trying to drop non-existent users or roles.

Now the above example is handled as follows:
When a user is dropped during handling the request, it causes a schema change
that leads to the rejection of the operation on the predicate check.
Upon retry, the initiator will detect that the user has been dropped
and handle it accordingly.