fix!: allow the plugin to be dropped leaving it's data in the database
Closes #996

This was the intended behavior all along, but I erroneously remove it
thinking that it's a bad thing that we're going to be having references
from _pico_plugin_migration to plugins which don't exist. If you drop
the plugin leaving it's data, the only way to drop that data would be
to re-create the same plugin and then drop it using `DROP PLUGIN WITH
DATA` syntax.

Also note that there's currently no way to drop the plugin's data
without dropping the plugin itself (other then from lua via
pico.migration_down but no analogue in SQL).