test: update log prefix asynchronously when instance name not known at startup
Previously in our integration tests we would explicitly provide instance
names on startup. One of the reasons for that was, that we process all
the logging output from the instances by adding the prefix with the
instance name to them. After we stopped generating names explicitly this
feature broke, and the logs started being prepended with just the listen
port, which is not as useful.

This commit partially resolves this issue by making it so that logging
prefix is automatically updated after the instance is bootstrapped and
we find out it's name. This is implemented just by checking the
Instance.name member variable, which is updated when
Instance.instance_info() method is called. It is implicitly called in
Instance.wait_online() which happens in almost every test (often
implicitly as well). But if you notice that your test output doesn't
contain instance names at the start of the lines, you may want to call
Instance.instance_info() explicitly at some point.

Also note that there's an inherent problem with that approach -- any
log output which is handled before the instance name is known will be
prefixed with just the port number. Hopefully this is not a very big
problem in most cases...

A better solution would be to add the instance_name to all the logger
output automatically from picodata.
See https://git.picodata.io/core/picodata/-/issues/1303