feat: rename plugin_dir -> share_dir
The parameter with the old name still works, but the deprecation warning
is reported upon use. Will be removed a bit later.

Even though we're technically allowed to break backwards compatibility
I've done it this way, so that we have some examples of how to do such a
thing for future reference.