refactor: simplify generic iterator code in plugins a bit
I still don't like this version, because we now have at least 3
instantiations of stop_background_jobs, which is pretty massive. The
better solution would be to collect the iterators into a Vec and pass
the slice to the function, which would make it so the final binary only
has 1 version of it. But the source code is a little bit more readable
in the current version, and considering how nobody cares that picodata's
release binary weighs 166MB...