fix: sentinel used to not downgrade instances after leader change
Originally we made it so sentinel only considers instances to be
Unreachable if there was a successful communication attempt at least
once before. This was done this way so that we don't stall bootstrap, as
newly added instances don't start responding to raft RPCs immediately.
But since then we hooked up proc_raft_interact on the receiving end to
the ReachabilityManager, so that if an instance sends us an RPC we
reset it's fail streak. After that change we now can be more eager about
downgrading unresponsive instances.

Closes #1369