test: cleanup test_migration_lock
Use SQL plugin API.

Use python threads for asynchronous tasks instead of lua eval with
tarantool fibers.

Note that pythons threads are very similar to fibers, in that they don't
work in parallel, instead only one thread can execute at a time while
the others are waiting for IO. So for example if one of the threads does
an RPC which blocks for some time other threads can do their work in the
mean time.