ci: enable tag fetching on test stage
Tag fetching is needed for our backwards compatibility tests to work,
because they rely on correct git tag information and git describe.
Unfortunately, this requires us to fetch all tags to the root of the
repository, which is Picodata itself, so that's may slow down whole
CI/CD pipeline running times, and may be fixed in future.