fix: snapshot generation script assumed any current tag as the latest one
If we're on 24.7.0, but `git tag` says there are more fresh 25.1.1, it
will create a directory for 25.1.1, even though it puts there a 24.7.0
recovery files. Also, it produced incorrect cluster name and the port
for PostgreSQL server protocol, which is not acceptable for tests on
backwards compatibility check, as `picodata run` without config waits
for the same name and protocol port by default.