feat: add activation_deadline parameter in the config file
This patch allows to configure parameter activation_deadline
for each instance in the configuration yaml.
By default, the value set to 7200 sec (2 hours)
which defines the maximum time that instance
will wait to be activated and join a cluster.
if the activation doesn't occur within this period
the instance will shut down.