build(.cargo/config.toml): add dynamic_lookup for x86
MacOS uses ld64 as linker and fails to link dylib
files if some of the required symbols are undefined.
That causes "testplug" (plugin stub for test)
build failure.

Though undefined symbols are expected, because
plugins will only be able to access them after
loading them to Picodata's process.

A workaround for this issue is to set linker policy
on undefined to "dynamic_lookup" and it has been
already done for ARM targets. This patch copies
that behaviour for Mac x86.