feat: add 'tier' field to governor steps Plan::Downgrade and Plan::ToOnline
This commit introduces a new 'tier' field
to the governor steps Plan::Downgrade and Plan::ToOnline.
The tier name is used in the metrics to report instance state.
These steps are used to update instance state (Online/Offline/Expelled).
The tier name represents the tier to which this instance belongs to.
Also this commit renames the field `target` to `instance_name`
in the struct Plan::ToOnline to reduce the tech debt