fix!: do not dump panic backtraces by default, log error if dump fails
Now, we won't produce `picodata-PID.backtrace` files by default. If you
want to dump panic backtraces, set `PICODATA_INTERNAL_BACKTRACE_DUMP`,
it is automatically set in our tests, though. Also, before that commit,
if we have panicked with some strange condition, and cannot dump the
backtrace to a file (e.g. out of disk space), the file writing error
was just ignored. Now, the error is logged and printed to the user.

This is technically a breaking change, because the observable behavior
changes, but this was never documented, so let's remove it silently.

Closes #1661