fix: make plugin cas to invoke wait_index
Before that all users of the API had to do that that on their end.
Besides that wait_index usage in that case was error prone since for
correctness users had to manually check term of the applied record and
in case it doesnt match with term returned from cas - whole operation needs
to be retried because that means that new leader accepted another
operation with this index.