fix(rpc): governor calls old rpc to enable plugins
Right after the release of 25.2.0 (8dbc7434), we have broken the
backwards compatibility by renaming `proc_enable_all_plugins` to
`proc_before_online`. Some time passed (1ba919e1), and we've returned
it back, while also redirecting it to a new implementation. Now, we
have to make governor use the old RPC, so that we do not break the
compatibility anymore.

Closes #1433.
Follow-up for !1662.