fix(sql): prevent panic in NamedWindows when using subqueries
We introduced a new entity, `referenced_named_window_ids`, to explicitly track which named windows are actually referenced under projection. This change was necessary because previously, if a query defined named windows but none of them were referenced in the projection, it could result in issues during the local SQL generation stage—specifically, an uninitialized parent in the output node.

Now, we filter out unused named windows and ensure they are not propagated to the local sql plan. If no named windows are referenced in the projection, the entire `WINDOW` clause is removed from the query.