fix: get_tags.py script now properly fails
Basically this is a cleanup of get_tags.py script.
Previously the script didnt fail when all attempts to fetch the tag were
unsuccessful. This led to unhelpful errors during the build stage. So
now fetch script will properly fail with relevant error message so it
will be easier to diagnose proper error.