feat(cli): introduce simple demo scenario
This commit adds a new command to our CLI - `picodata demo`. It starts a
cluster of instances with PGPROTO server enabled. After initializing the
cluster, it starts the PostgreSQL interactive terminal (`psql`), so the
user is ready to run SQL queries in it and see the changes they make.
This is useful feature both for the developers, and customers to show
something quickly and interactive, especially on conferences.

Addresses: #1967.