fix: don't use fiber::reschedule in pico.raft_log()
The use of reschedule in that case is erroneous, because the idea was to
keep the event loop going, but that is not what reschedule does.
Reschedule would just lead to an infinite loop with extra steps because
it skips the epoll_wait call, thus making it impossible for IO events to
be noticed by other fibers.

Using fiber::reschedule inside a loop {} is almost definitely incorrect!