fix: upgrading from 25.1.1 to 25.2.1 was broken
We added new stored procedures `.proc_instance_uuid` && `.proc_before_online`
and that broke our storage bootstrap routine. This commit fixes the
issue in that now it is safe to run a newer version of picodata on
snapshots generated by an older version.

This fix was tested manually for launching picodata-25.2.2-ish on 25.1.1
snapshots. Integration test is not added as it's a whole separate issue
in #1037.

Note that the clusters which were broken by this bug will not be
automatically fixed and will require some manual actions.

Note also that storage schema will not be upgraded if newer version of
picodata is launched on older snapshots. That feature is tracked in #1678.