fix: don't delete from _cluster if not in _pico_instance
Previously we made it so instances are removed from _cluster if they're
not in _pico_instance. This was a mistake because that code was
executing as part of .proc_replication, which may be called with lagging
global state, which means that instances are added to _cluster before
_pico_instance.

So for now we don't remove records from _cluster if uuid not found in
_pico_instance, which is a potential problem as removing records from
from _pico_instance may break replication within a replicaset.
This problem should be addressed as part of #1110.