fix(sql): index creation fetched outdated space id
CREATE INDEX fetched outdated space ID from the space cache (that
actually doesn't work for picodata). As a result, in a case of ABA
for spaces with different amount of indexes we could easily end with
panic on insertion to _pico_index.
As a solution, we stopped using space cache (as we do not use it
anywhere else in the picodata code) and switch to Space::find.