fix: sentinel backoff algorithm used to use incorrect raft index
When doing the so called raft read barrier it is very important to
choose the correct raft log index. In this case we must remember the
applied index at the moment the request is constructed, because that
index corresponds to the state of global tables we were looking at.
Before this commit the behavior was wrong: we took the applied index
after sending an RPC request during which our raft state machine had a
good chance of applying new raft log entries including ones affecting
the global tables we were looking at (_pico_instance). So there was a
theoretical scenario where sentinel would block attempting to do a raft
read barrier even though the raft log has already been advanced