chore: ditch cargo proxy setup because of problems with crate publishing
The problem is that despite having proxy we still want to publish to
original crates.io registry. With proxy default publish routine stops
working. There is a workaround using `--registry` argument of cargo
publish set to `crates-io` which is a default value (?). This workaround
falls short with case when several dependant crates need to be
published. For example we have picodata-plugin-proc-macro and
picodata-plugin which depends on picodata-plugin-proc-macro. So when we
publish proc macro crate it succeeds but publication of main plugin
crate fails because cargo cant find new version of proc macro crate in
proxy registry because it havent propagated yet from crates.io.

Removal of proxy shouldnt harm our CI process because we cache cargo
registry between builds. So we only should reach to cargo when
Cargo.lock changes.