feat: make it so raft heartbeats include applied index
This will be used in a future commit to implement detection of instances
with failed raft_main_loop. The idea is, we can easily notice that
instance is broken if it's applied index is lagging behind the leader
and is not progressing. Such instances harm the cluster as for example
a call to proc_sharding will always timeout on them.

Note that this commit introduces a change into the request encoding for
the `.proc_raft_interact` stored procedure and does so in a backwards
compatible way. Such that picodata versions before and after this commit
will be able to successfully communicate.