feat: extend proc_raft_interact messages with extra info
Add the following fields to the raft messages:
- version: numeric value describing the format of the message itself
- flags: miscellaneous bitflags (will be used in a following commit)
- applied_index: will be used in future to detect degraded instances.

The idea for applied_index is, we can easily notice that instance is
broken if it's applied index is lagging behind the leader and is not
progressing. Such instances harm the cluster as for example a call to
proc_sharding will always timeout on them.

Note that this commit introduces a change into the request encoding for
the `.proc_raft_interact` stored procedure and does so in a backwards
compatible way. Such that picodata versions before and after this commit
will be able to successfully communicate.