fix: make DDL operations handle log compaction
Previously, the DDL command code could not handle log compaction.
The catch is that, to determine the outcome of an operation
(commit/abort), a DDL command inspects its local raft log. But that
log can be compacted at any time by the raft leader over RPC. As a
result, the DDL command may never see whether it committed or aborted.
However, in many cases it can infer the outcome by checking whether
the operation was applied. This commit adds exactly that inference
logic.