dev: add a dummy temp project to the workspace
This project is useful when you want to quickly through some code and
check if it compiles, runs. You could also just through your code into
picodata's main.rs, but it will force a full rebuild which for picodata
is massive due to the large amount of dependencies including tarantool-sys.