ci: fix staged diff
$MAIN_BRANCH should use a push cache policy so that master doesn't pull the previous master.
This is important because there may be uncommitted changes when using git fetch -uf

dev branches should use a pull-push cache policy so that tests have the corresponding dev branch
that we fetched during the clone step. This is important because tests should check the dev branch, not the master.