test(compat): fix snapshot patch version check constraint
In our current setup, versioning does not start from PATCH part of
version zero (`0`) as per SemVer, but from one (`1`). Because of this,
the backward compatibility test for the PATCH version part would fail
right after a new release. This commit fixes the issue by detecting such
case and skipping the test when applicable.