fix(compatibility): snapshot generation script performed unnecessary path checks
Currently, we check for path existence and if this path has some files
in it for compatibility tests to gather previous version snapshot files.
But, as long as we are generating a snapshot for further testing, we
don't need these checks as it is expected for paths to be non-existent
and be empty, respectively.