test(rolling): introduce rolling upgrade facility
In our current CI we can’t test different versions of Picodata. We need
a "moving" upgrade coverage, which we can apply to every major Picodata
release. However, each release ships its own incompatibilities and
new features, so we need a way to abstract these out. This is why we
need a custom testing framework where we implement upgrade scenarios
and topology operations in heterogeneous clusters and, using cluster
and separate instances information, determine whether a scenario is
successful or not.

Closes: #1037.