fix(log): do not warn unsecured ldap connection
IPROTO authentication implementation does not send an information about
method used for the specified user (e.g., MD5) to the client. That
means, we have to try every possible method by hand, including LDAP. If
the user does not use LDAP, warning about unsecured attempt to connect
will still be printed, which is not expected behavior, even if LDAP will
be the last method used in enumeration (i.e., invalid password still
need to try all methods). In addition, if we run out of attempts to
enter the password, the user may be blocked.

For more context, look at PostgreSQL authentication protocol, which
sends the appropriate authentication method for the specified user, and
does not require a client to iterate through all available methods.

Note that this change will only affect `picodata` feature flag.

Addresses: !1833, #1734.