rename: Catalog::{tables -> pico_table} & storage::{Tables -> PicoTable}
From now on all structs for accessing system tables should be named the
same as the tables are named in the storage engine. This commit changes
naming for `_pico_table` related structs, i.e.
- `Catalog::tables` is renamed to `Catalog::pico_table`
- `storage::Tables` is renamed to `storage::PicoTable`

Historically this wasn't so, because when first system tables were added
to picodata they were named differently, for example `box.space.instance`
instead of `box.space._pico_instance`. And in that time it was confusing
for the table-access struct to be named simply `Instance`, because that
would confict with struct `Instance` which describes the row of that
table. But since then we changed the naming scheme for system tables and
now it makes more sense to have source-code level names matching the
schema-level names.