fix: don't set read_only = false in proc_sharding during master switchover
This fixes a bug in master transitioning logic. The idea is when we're
switching over a replicaset's master the new master must first
synchronize with the old one (assuming the old master is Online for
that).

For the duration of this sync both instances should be read_only, so the
possibility to introduce a replication conflict is reduced. This logic
was previously broken because current_master would do
`box.cfg { read_only = false }` implicitly when configuring vshard
storage.