feat: add SubQueryReference node to IR
In this commit we add a SubQueryReference node to the IR.
The main reason is to get rid of the is_additional_child function usage.
The only place where we add it to the IR is Plan::add_replaced_subquery.
Adding the SubQueryReference node type is also a step towards excluding
the array of children from relational nodes in the IR.