perf: don't use RegionBuffer for Response::encode_rmp
I did some basic benchmarking and it turns out that serializing into the
rust Vec and copying onto the region afterwards is faster than
serializing onto the region. If I had to guess this has to do with the
fact that working with the region allocator requires passing through the
ABI bounday, which is not as chip as it may seem.

Not deleting the RegionBuffer entirely for now, because it's exported
publicly in the SDK.

(cherry picked from commit b9eb8c2ee568a2e60ffe64020c880591eb9bd1f8)