test: display info about governor status when Instance.wait_online fails
Previously when `Instance.wait_online()` of `Cluster.wait_online()`
would fail we would see somethings like this:

```
AssertionError:
assert 'Offline' == 'Online'

  - Online
  ?  ^
  + Offline
  ?  ^^
```

And now the output will be more like this:
```
AssertionError:
  Timed out waiting for instance 'default_1_1' state 'Online'.
  Expected state 'Online', actual: 'Offline'
  Instance log file: /tmp/pytest-of-user/pytest-1234/tmp0/3301/picodata.log

  Meanwhile governor is 'default_1_1'
  Governor loop status is: 'update instance state to online'

  Last governor error is:
  ---
  code: 78
  file: src/governor/mod.rs
  line: 787
  message: timeout
  ...

assert 'Offline' == 'Online'

  - Online
  ?  ^
  + Offline
  ?  ^^
```

So now the error message contains name of instance which didn't become
Online in time. It also contains info about governor: which instance is
currently governor (so you can go look at it's logs), what was the
governor trying to do and what was the last error that it encountered.