docs(changelog): auth method flags default values clarification
In !1833 (@aa78cf34), we introduced a mechanism to authenticate users
without requiring an explicit authentication method. However, the
changelog entry in that commit described an outdated version of the
change. In reality, we do not ignore authentication method flags when
they are provided. Instead, we avoid applying default values to these
flags when they are omitted, and rely on the newly introduced mechanism.