feat: migrate SQL executor to port API
Previously we packed the entire query result into a single tarantool
tuple and then wrote it to the port, which was inefficient and caused
numerous data repackings in rust memory. Now we explicitly pass the
tarantool port into the internal SQL execution API and stream data
into it as execution proceeds, then dump to `obuf` via the port’s
virtual-method logic.

Co-Authored-By: ai <a.yurin@picodata.io>
Co-Authored-By: Erik Khamitov <e.khamitov@picodata.io>