config: increase pg_statement_max/pg_portal_max to 100 000
Picodata is an in-memory database and typically maintains
thousands if not tens of thousands client PostgreSQL connections.

Since each connection effectively has its own prepared statement
cache the default setting for 1024 statements is amplified
by the total number of connections.

A complete removal of the limit seems dangerous, since it
allows to identify and protects from client applications which leak
prepared statement handles.