fix: copy proc_rpc_dispatch arguments to rust memory
This should fix the problem showcased in https://git.picodata.io/picodata/plugin/sirin/-/issues/130.
The original problem presumably goes something like this:
- tarantool receives an iproto request via network, the request is
  stored in an ibuf struct, owned by the connection object
- tarantool calls our stored procedure proc_rpc_dispatch, passes a
  pointer into that ibuf struct
- proc_rpc_dispatch is doing some work and yields for a long time
- tarantool receives another request on the same connection, and the
  ibuf buffer is overwritten with the new data
- original proc_rpc_dispatch fiber wakes up and reads the arguments via
  the old pointer which now points to new data

Unfortunately I wasn't able to confirm or deny this conjecture yet. But
if it's true, then the fix in this commit should solve the problem, as
we will always copy the data into the buffer controlled only by us.