fix(audit): ignore fxhash and remove async-std
Suddenly, `async-std` has been deprecated, so previously stated `slab`
transitive dependency is no more needed to be stated, as long as the
whole `async-std` is unmaintaned anymore. As of the `fxhash`, it was
practically abandoned for 5 years already, but in the `raft-rs` crate,
there were no issues of using it, so we just ignore this warning.