fix: don't skip raft snapshots with stale schema version
It looks like this behavior was always a mistake. schema_version should
have never been used to determine raft snapshot actuality, this is what
the raft index is for.

It's a good time to note that it's possible that global schema is going
to be lagging behind shard-local schema, i.e. a table can be in _space
but not it _pico_table for some period of time. But not the other way
round!

For instance something like this could happen:
- `replica` requests a raft snapshot from `raft_leader`
- `raft_leader` creates raft snapshot, puts current `global_schema_version`=1
  as the snapshot's schema version
- `raft_leader` sends raft snapshot `replica`
- a new schema change is applied, `global_schema_version` is increased to 2
- `master` from same replicaset as `replica` applies the schema change
- `master` sends the shard-local schema changes (e.g. `_space` contents
  as well as `_schema.local_schema_version`=2) to `replica` via
  tarantool replication stream
- `replica` receives shard-local schema changes, has `local_schema_version`=2
- `replica` receives raft snapshot, and sees snapshot schema version (=1) <
  local_schema_version (=2)

This can happen due to network effects considering that raft log
replication and tarantool replication are transferred using separate
TCP connections.