fix: TRUNCATE used to not work correctly for global tables
Previously the implementation didn't take into account global tables. As
a result when truncating global tables only the master replicas would
apply the operations and all of the replicas would simply ignore it.
See the test in this commit, it used to fail before.

This fix makes it so we apply TRUNCATE operations differently based on
the distribution of the table. If the table is global governor will not
send any RPC to masters and instead propose the DdlCommit raft operation
immediately. And when applying the DdlCommit operation in raft_main_loop
the operation is always applied no matter if we're on the master or a
read-only replica.

(cherry picked from commit 58fbbeb9a2f4d9a5cd931a81ded73281131d9ffc)