fix: new ALTER SYSTEM parameter plugin_check_migration_hash
Previously we would not allow migration files to be changed after the
plugin was created. This was done intentionally to avoid situations
where migration files are changed by mistake which could cause hard
to debug errors. But in practice this restriction turned out to be
more harmful than expected, because it makes it much more difficult for
plugin developers to fix errors in migration files.

So now there's a parameter which disables this validation. If
plugin_check_migration_hash = false then migration file checksum
mismatch will only cause a warning to be logged and the command will
proceed with the new migration files.