test: automatically create symlinks for instance directories
Before this we would usually create instance directories named after
the listen ports of instances, because instance names would not be known
at the moment instance directory was designated. Now after we find out
the instance's name we create a symlink to the instance directory with
that name for our convenience.

For example now you can lookup instance directories by these paths:

    /tmp/pytest-of-<username>/pytest-current/tmpcurrent/<instance_name>