test: Cluster.add_instance parameters for logging into a file
Add 2 new parameters to Cluster.add_instance():
    - log_to_console (default True)
    - log_to_file (default File)

If both are true then picodata will be started with this environment variable:

    PICODATA_LOG="|tee ${INSTANCE_DIR}/picodata.log"

If only `log_to_file` is True then it will be like this:

    PICODATA_LOG="${INSTANCE_DIR}/picodata.log"

If none of them are True then logs are disable this way:

    PICODATA_LOG="/dev/null"

The default behavior didn't change. By default the logs are still only
printed to stderr.

If `log_to_file=True` then logs are written to `picodata.log` file in
hte instance directory. You can find them after the test finished for
example like this:

    /tmp/pytest-of-<username>/pytest-current/tmpcurrent/<instance_name>/picodata.log