fix: automatically retry retriable CAS errors in global DML
Before this fix when doing DML on global tables if a retiable CAS error
happened we would return this error to the user. For example if raft log
got compacted or a conflicting DML from a concurrent client arrived, the
operation would fail.

Now we automatically retry such errors so users should never encounter
RaftLogCompacted errors when running SQL commands.

ExecutionPlan:take_subtree will now not mutate the plan in case of DML
on global tables, instead a copy of the subtree is made.

Close #2273