fix: raft snapshots used to contain incorrect term sometimes
Before this fix we would always put the current term into the raft
snapshot metadata when sending it to the catching up instance. This
meant that sometimes the raft term in the snapshot would not correspond
to the raft index in snapshot.

This could happen for example if a raft snapshot was generated in a
period of time after the leader change (term increase) and before the
first entry in the new term was applied (hence the applied entry's term
would not match the current term).

As a result the instance which received such snapshot would corrupt it's
metadata, and reject any subsequent append entry messages from leader
because of the entry index + term mismatch.

We fix this by persisting the `applied_term` along with the `applied`
index and using `applied_term` instead of `term` for the raft snapshot
metadata.

(cherry picked from commit ee0f15a5ac741298955bc09446fea069a447a657)