refactor: make ddl abort handling similar to ddl commit wrt local schema version
Clarify the logic behind when we do need and when we don't need to
update the local schema version.

Basically the complexity lies in how our DROP operations are implemented.
When applying DROP operation the local schema version is changed when we
apply the DdlCommit raft op, whereas for other DDL operations the
version is changed during proc_apply_schema_change RPC from governor.

As a result when aborting the DROP operations the local schema version
doesn't need to be changed, because it wasn't bumped yet at that point.

Also add logging whenever we change local schema version for debug
purposes