fix: send explicit is_master flag in response to proc_raft_join requests
Previously when joining new instances to the cluster we would only
return a list of replica addresses, which would be used both for the
`box.cfg { replication }` parameter as is, and for `the box.cfg { read_only }`
parameter (`true` iff `replication.len() == 1`). This used to work well,
because topology updates go through raft log, which makes them
linearizable, and hence at the moment the first replica is joining the
`replication` array always has length 1, and when any other replica is
joining the length is > 1.

However this changed after we made it so proc_raft_join became idempotent.
Now it's possible for something like his to happen:

- storage_1_1 -> RPC proc_raft_join -> { replication = storage_1_1 }
- storage_1_1 (CRASHES)
- storage_1_2 -> RPC proc_raft_join -> { replication = storage_1_1, storage_1_2 }
- storage_1_1 (RESTARTS) -> RPC proc_raft_join -> { replication = storage_1_1, storage_1_2 }

This means that using the `replication` array to determine the value for
the `read_only` parameter is no longer viable. Hence we add a new flag
to .proc_raft_join return packet.

Note that this change is potentially incompatible with previous versions
of picodata, however we do not support the case of joining new instances
to a cluster of multiple different versions, so it shouldn't be possible
for anybody to encounter compatibility problems due to this change.