fix: used to panic sometimes when joining the cluster
When booting up the instance we check if we're the only voter and in
this case we attempt to promote to leader and even assert that the
attempt was successful. The problem is there's a case where this could
fail even though we make an attempt to apply any configuration changes
there still may be one in progress which we don't see.

The simplest solution is to just remove the assert. This shouldn't be
a problem for us, because after this the normal raft leader elections
should take place.