test: reduce bucket rebalancing flakiness in tests
I think I finally figured out the problem we were having with tests
which were failing in `wait_until_instance_has_this_many_active_buckets`.

Bucket rebalancing may get stuck for a while if an replicaset is joined
while buckets are being transferred. After that the recovery fibers are
responsible for the cleanup and rebalancing will not happen until the
cleanup is finished.

So we must wake up recovery fibers on every instance to speed up the
rebalancing in such cases.

We also wake up the gc fibers just in case.

Close #1864
Close #1422
Close #1397
Close #807