fix: instance state metric to report correct values on all instances
Before this patch pico_instance_state was only recorded on raft leader.
It was misbehaving in several ways.

First, we didnt clean previous value when instance state has changed.
Since we report state as a label the value with old label was not removed,
so for example both Online and Offline label values could be seen.

Second, since the value was recorded only on leader (in governor)
leadership transition didnt remove old values. So conflicting entries
could be reported from new and old leader.

Solution to the first problems is to remove previous values before
recording new one. Second problem is solved by reposting metric from all
instances. For that to work correctly metric is refreshed in the
following cases:

1) when applying raft record that changes instance current_state
2) when instance starts up in postjoin
3) when raft snapshot is applied