fix: raft_wal_count_max default value changed from 64 to 16384
The original default of 64 was chosen arbitrarily without any good
thought being put into it. In practice this value turned out to be bad,
because frequent DML to global tables would result in frequent log
compaction which causes need for relatively expensive raft snapshots.

In the mean time it's actually not obvious that we need to limit the
number of raft log entries at all, because we already have
`raft_wal_size_max` which limits the size of raft log in bytes and
defaults to 64MiB, so removing raft_wal_count_max entirely would make
it so the actual cap would be dynamic based on the size of actual raft
log entries.

Practical use for non infinite raft_wal_count_max is found in tests and
debugging, as it allows to trigger log compaction more reliably compared
to raft_wal_size_max.